static repo  bug 修复...
